fix: make it work with the new structure
