+++ src/commands/utils/config/default_values.rs
@@ -6,10 +6,10 @@ pub const DEFAULT_L2_EXPLORER_URL: &str = "http://localhost:3010";
 pub const DEFAULT_L1_EXPLORER_URL: &str = "";
 pub const DEFAULT_PRIVATE_KEY: &str =
     "0x850683b40d4a740aa6e745f889a6fdc8327be76e122f5aba645a5b02d0248db8";
-// 0x36615Cf349d7F6344891B1e7CA7C72883F5dc049
+// 0x5e6d086f5ec079adff4fb3774cdf3e8d6a34f7e9
 pub const DEFAULT_ADDRESS: H160 = H160([
-    0x36, 0x61, 0x5C, 0xf3, 0x49, 0xd7, 0xf6, 0x34, 0x48, 0x91, 0xb1, 0xe7, 0xca, 0x7c, 0x72, 0x88,
-    0x3f, 0x5d, 0xc0, 0x48,
+    0x5e, 0x6d, 0x08, 0x6f, 0x5e, 0xc0, 0x79, 0xad, 0xff, 0x4f, 0xb3, 0x77, 0x4c, 0xdf, 0x3e, 0x8d,
+    0x6a, 0x34, 0xf7, 0xe9,
 ]);
 // 0x5E6D086F5eC079ADFF4FB3774CDf3e8D6a34F7E9
 pub const DEFAULT_CONTRACT_ADDRESS: H160 = H160([
